move simulations into package
